DEVENGAGE-1986: Cleaned up and documented code.
